==> stackoverflow/preproc/00 descriptivos para ordenar.R <==
source("init.R")

zz=gzfile('data/input/StackOverflow/developer_survey_2018/survey_results_public.csv.gz','rt')   
dataset_raw = read.csv(zz,header = T)

interaction = function(x){
  dataset_raw %>% group_by_at(vars(one_of(x))) %>% summarise(salary = mean(as.numeric(ConvertedSalary),na.rm = T),
                                                             numero = n())
}

interaction("Gender")

table(dataset_raw$Hobby)
table(dataset_raw$OpenSource)
table(dataset_raw$Student)
table(dataset_raw$Employment)
table(dataset_raw$FormalEducation)
table(dataset_raw$UndergradMajor)
table(dataset_raw$CompanySize)
table(dataset_raw$DevType)
table(dataset_raw$YearsCoding)
table(dataset_raw$YearsCodingProf)
table(dataset_raw$JobSatisfaction)
table(dataset_raw$CareerSatisfaction)
table(dataset_raw$HopeFiveYears)
table(dataset_raw$JobSearchStatus)
table(dataset_raw$LastNewJob)
table(dataset_raw$Currency)
table(dataset_raw$Salary)
table(dataset_raw$SalaryType)
table(dataset_raw$ConvertedSalary)
table(dataset_raw$CommunicationTools)
table(dataset_raw$TimeFullyProductive)
table(dataset_raw$EducationTypes)
table(dataset_raw$SelfTaughtTypes)
table(dataset_raw$TimeAfterBootcamp)
table(dataset_raw$HackathonReasons)
table(dataset_raw$AgreeDisagree1)
table(dataset_raw$AgreeDisagree2)
table(dataset_raw$AgreeDisagree3)
table(dataset_raw$LanguageWorkedWith)
table(dataset_raw$LanguageDesireNextYear)
table(dataset_raw$UpdateCV)
table(dataset_raw$IDE)
table(dataset_raw$OperatingSystem)
table(dataset_raw$NumberMonitors)
table(dataset_raw$Methodology)
table(dataset_raw$VersionControl)
table(dataset_raw$CheckInCode)
table(dataset_raw$AdBlocker)
table(dataset_raw$AdBlockerDisable)
table(dataset_raw$AdBlockerReasons)
table(dataset_raw$AdsAgreeDisagree1)
table(dataset_raw$AdsPriorities1)
table(dataset_raw$AIDangerous)
table(dataset_raw$AIInteresting)
table(dataset_raw$AIResponsible)
table(dataset_raw$AIFuture)
table(dataset_raw$EthicsChoice)
table(dataset_raw$EthicsReport)
table(dataset_raw$EthicalImplications)
table(dataset_raw$StackOverflowRecommend)
table(dataset_raw$StackOverflowVisit)
table(dataset_raw$StackOverflowHasAccount)
table(dataset_raw$StackOverflowParticipate)
table(dataset_raw$StackOverflowJobs)
table(dataset_raw$StackOverflowDevStory)
table(dataset_raw$StackOverflowJobsRecommend)
table(dataset_raw$StackOverflowConsiderMember)
table(dataset_raw$HypotheticalTools1)
table(dataset_raw$WakeTime)
table(dataset_raw$HoursComputer)
table(dataset_raw$HoursOutside)
table(dataset_raw$SkipMeals)
table(dataset_raw$ErgonomicDevices)
table(dataset_raw$Exercise)
table(dataset_raw$Gender)
table(dataset_raw$SexualOrientation)
table(dataset_raw$EducationParents)
table(dataset_raw$RaceEthnicity)
table(dataset_raw$Age)
table(dataset_raw$Dependents)
table(dataset_raw$MilitaryUS)
table(dataset_raw$SurveyTooLong)
table(dataset_raw$SurveyEasy)
